fix: converted i32 to f32

+++ src/temperature_converter.rs
@@ -8,7 +8,7 @@ pub fn temperature_converter_cel_to_fah() {
 
    io::stdin().read_line(&mut room_temperature).expect("Could not read from standard input");
 
-   let room_temperature:i32 = match room_temperature.trim().parse() {
+   let room_temperature:f32 = match room_temperature.trim().parse() {
         Ok(num)=>num,
         Err(_)=>panic!("The temperature could not be converted to a number")
    };
@@ -18,9 +18,9 @@ pub fn temperature_converter_cel_to_fah() {
 }
 
 
-fn convert_celsius_to_fahrenheit(temperature:i32)-> i32{
+fn convert_celsius_to_fahrenheit(temperature:f32)-> f32{
 
-   let converted_temperature = 5.0/9.0 * ((temperature) - 32);
+   let converted_temperature = 5.0/9.0 * (temperature - 32.0);
 
    converted_temperature
